feat(robot-server): Use a BadRun when we cant load
Up to now, if there's a run saved in the persistence layer that cannot
be loaded - which is typically because it contains data from a version
of the robot server or api package that whatever's currently running
can't handle - we error when trying to retrieve it. That includes both a
500 error when trying to access that particular run, which clients can
broadly handle, and a 500 error when trying to list all runs, which
clients cannot. Without being able to list out all runs, there's no way
for clients to find the problematic run - the IDs are UUIDs and cannot
be enumerated - and remove it. The only recourse is to delete all the
run storage.

A different way to handle this problem is to consider a "bad run", a run
whose run metadata or engine state summary cannot be loaded, as a first
class entity that can be returned from run access endpoints wherever a
run could be, without an HTTP level error. This is done everywhere for
consistency in this commit, though the argument could be made that it
should only be done in the list-all-runs access and other endpoints
should continue to error.

This bad run contains error information about the cause of the invalid
data using a new enumerated error. The bad run will carry all the
information that could be loaded - in effect, if the state summary is
bad then the run metadata will still be present, and the ID should
generally be accessible.

Closes EXEC-344