feat(app): Display bad runs on the ODD (#14736)
When runs are bad, which means they couldn't be loaded correctly, we now
get standin records called BadRuns rather than 500 errors. We should
display that a run is bad on the ODD. For now, let's try making it red
and displaying a status text.

This is what this looks like. Thoughts?

![image](https://github.com/Opentrons/opentrons/assets/3091648/11da4921-5901-4157-86ce-5e45a79cac26)