feat(app): add tooltips to ChooseRobotToRunProtocolSlideout RTPs
According to designs, number type RTPs should display their respective numerical ranges as
InputField captions. The optional RTP description should be displayed in a tooltip when an
information icon is hovered next to the RTP title. As part of the scope of this change, I also
refactor the runtime parameter types so that the type of the RTP (numerical vs choice vs boolean)
will narrow to the appropriate properties of that specific RTP when mapping.