feat(api-client,app,react-api-client): implement ODD anonymous localization provider
implements a localization provider for the ODD app that substitutes anonymous translation values
when the "enableOEMMode" robot setting is on. pushes the localization provider further down the DOM
to be within the ODD api host provider for the robot settings request, and splits out a separate
instance for the desktop app. refactors the OnDeviceDisplayApp file a bit to avoid unnecessary
rerenders of providers - currently the entire app rerenders on every route change because of the
scroll ref. adds api client and react api client functions for robot settings get requests.

closes PLAT-265