chore(api): fix typing of task_queue workers (#14755)
task_queue has these cleanup and run functions that you pass in with
some arguments to bind. None of this was typesafe because it predated
ParamSpec. Now that we have ParamSpec, we can make these typesafe.

## changelog
typing only in `task_queue`. `set_run_func` and `set_cleanup_func` get
paramspecs. `set_run_func` is trivial; `set_cleanup_func` has a
requirement to have an `error` argument (which has to be the first
argument, for "paramspec still isn't good enough" reasons, but that was
already an implicit requirement).

also, we do a manual closure instead of `partial()` because I'm pretty
sure `partial()` isn't actually typesafe.