feat(app): mark protocol anlayses as stale if they lack RTP
We currently display a stale anlaysis banner for protocols that are determined to require
reanalysis. The current check is for the lack of a `liquids` key on the analysis output. Now, we
also want to check for the lack of a `runTimeParmeters` key, as all analyses should contain this
key, even if its value is an empty list.