feat(api-client,app,react-api-client): implement ODD anonymous localization provider (#14741)
implements a localization provider for the ODD app that substitutes
anonymous translation values when the "enableOEMMode" robot setting is
on. pushes the localization provider further down the DOM to be within
the ODD api host provider for the robot settings request, and splits out
a separate instance for the desktop app.

refactors the `OnDeviceDisplayApp` component a bit to avoid rerenders of
providers - currently the entire ODD app rerenders on every route change
because of the scroll ref.

refactors the initial loading screen out of the route tree and up the
DOM to block localization provider render until the robot-server api is
up

adds api client and react api client functions for robot settings get
requests.

all translation keys that reference "opentrons" or "flex" are moved to
the new `branded.json` and `anonymous.json` files. anonymous copy will be finalized in PLAT-243

closes PLAT-265