feat(performance-metrics): add RobotContextTracker (#14862)
# Overview

Create RobotContextTracker class.

This class provides a `track` method. 
The track method should be used as a decorator on either a
function/method. It takes a RobotContextState enum value to label what
state the RobotContextTracker is tracking. Uses `FunctionTimer` to
measure execution time and stores results in a list.

RobotContextTracker is defaulted not to track anything at all. To turn
on tracking, instantiate the class with `should_track=True`. When not
tracking, the `track` method calls the underlying wrapped function as
quickly as possible.

# Test Plan

- See test_robot_contest_tracker.py

# Changelog

- Add RobotContextTracker class
- Add test_robot_context_tracker.py

# Review requests

None

# Risk assessment

Low, not being used on any production code