feat(app): push recent RTP run card to protocol details rather than setup
Because we don't yet have designs displaying details of a recent protocol run card with specific
runtime parameter values, we want to allow the user to reset parameter values afer clicking a recent
run card for an RTP-containing protocol. Behavior for non-RTP protocols will not change and will
push the user directly to protocol run setup.