fix(app): set max height for desktop modal shell
We special-cased the size of the modal shell to accommodate a large body for the 96-channel pipette
attachment BeforeBeginning step to avoid scrollable content. The size was set to 70% of the screen,
but that resulted in a lot of modal whitespace when the app was resized to larger than default
dimensions. Here, I add a max height of 31rem to accommodate the content of the largest pipette
wizard flow tile content without growing arbitrarily large with a resized desktop app. I also ensure
minimum padding between tile content and footer buttons, and fix padding on the ChoosePipette tile.