feat(shared-data, api): add uiMaxFlowRate key to pipette definitions (#14859)
closes AUTH-328

This PR add max flow rate values to each GEN2 and GEN3 pipette model
definition to be used for FE applications.

Rather than having a max flow rate for every pipette's supported tip's
volume for aspirate and dispense, we have only 1 max flow rate for every
pipette's supported tip that is. That value is based off of the lowest
volume minus 2% to account for safety.

For example:
`p1000_single_v3.6` has a `minVolume` of 5uL. The `t50` tips has a max
flow rate at volume 5uL of `801.3`. So the `uiMaxFlowRate` value is 98%
of that which is `785.2`

🔈 Additionally, the `lowVolumeDefault` default flow rates for
`aspirate`, `dispense`, and `blowout` have been changed to match the
`uiMaxFlowRate`. This DOES NOT result in a physical change behavior and
the change is needed for ui purposes.