fix(app): reset robot and protocol slideout states on close
When a user clicks out of a ChooseRobot or ChooseProtocol slideout, the slideout should 1) reset to
the first page if a multi-page slideout, 2) reset default robot or protocol selection to the first
valid option if one exists, and 3) restore runtime parameter values to default if setting up
protocol. Note that if a user is on runtime parameters step and selects 'Change robot', any
overriden parameter values will persist when parameter setup screen is proceeded to again.