feat(api): implement labware.set_offset in 2.18
The python protocol api's labware.set_offset() command lets you set a
labware position check offset programmatically. This is useful in
protocols run outside the app (i.e. through Jupyter, through
opentrons_execute).

When we moved to the protocol engine, we didn't support running those
protocols outside the app, and therefore we didn't implement this method
on those new API versions.

Now, we allow that again, so implement that method again.

This relies on dispatching AddLabwareOffset actions at arbitrary times,
because it will set an offset for the location in which the labware is
_currently_ present. This allows the user to set offsets without having
to wrangle with the engine's internal definitions of where something is,
at the cost of the user having to spread their offset calls around their
protocol if they have protocols that load things in one place
and later move them to another.

Closes RSQ-29