fix(hardware): remove can messenger `motor_enabled` listener (#14943)
<!--
Thanks for taking the time to open a pull request! Please make sure
you've read the "Opening Pull Requests" section of our Contributing
Guide:


https://github.com/Opentrons/opentrons/blob/edge/CONTRIBUTING.md#opening-pull-requests

To ensure your code is reviewed quickly and thoroughly, please fill out
the sections below to the best of your ability!
-->

# Overview
When #14479 was added, the motor enable message listener was never
cleaned up properly and thus causing the robot to slow down
significantly as it runs more and more move commands.