Restore original `ProtocolEngine.estop()` flag checking.
This fixes problems I introduced in this PR where certain E-stops would still try to drop tips and home at the end of the run.

I thought that inspecting the `error` argument passed in to `finish()` would be enough to determine whether the run is finishing because of an E-stop. It's not. When the physical E-stop button is pressed, our top-down interruption from `ProtocolEngine.estop()` races against any bottom-up interruption from the electronics. The top-down interruption can either:

* Raise a `RunStoppedError` when the protocol tries to run a new command
* Raise a `RunStoppedError` when when the current command is asyncio `.cancel()`'d

Neither of which currently carry any sign that they came from an E-stop. Until they do, we need a separate `from_estop` flag in `CommandState`.