fix(api): enagage axis before attempting to move
There seems to be some issue on the firmware where the motor is enabled but does not seem to actually get enabled, causing the axis we are attempting to move to error with finished movement with condition not met. The firmware should be enabling the motor and if we query the motor enable status with get_status_request = 0x01 we get the correct response where the motor is enabled. For whatever reason sending an explicit enable_motor_request = 0x06 before moving the axis fixes the problem.