fix(api): engage axis to enable the motor before attempting to move the axis. (#14955)
There seems to be some issue on the firmware where the motor is enabled
but does not seem to get enabled, causing the axis we are attempting to
move to throw a `finished movement with a condition not met` error. The
firmware should be enabling the motor, and if we query the motor enable
status with get_status_request = 0x01 we get the correct response where
the motor is enabled. For whatever reason sending an explicit
enable_motor_request = 0x06 before moving the axis fixes the problem.

---------

Co-authored-by: ahiuchingau <20424172+ahiuchingau@users.noreply.github.com>