fix(app): fix infinitely re-rendering/never rendering firmware success toast
There are two issues with rendering firmware update toasts. First, the toasts depend on a request id
stored as state within the ModuleCard, but ModuleCards most often unrender during the firmware
update process, so this state is lost. This causes the toast never to render. The second issue is
that sometimes, given the timing of the polling for attached modules, the module is always attached
during the firmware update, thereby causing the module card not to unrender. When this happens, the
useEffect hook responsible for making the success toast has conditional logic that is always true
after an update, causing the toast to render infinitely. The solution to both these problems is to
lift the request id state out of the module card itself and then abstracta way the storage/retrieval
via a utility hook, which is utilized by all parent components of ModuleCard.