Correct HS Time Calculation and Print Statements (#15005)
<!--
Thanks for taking the time to open a pull request! Please make sure
you've read the "Opening Pull Requests" section of our Contributing
Guide:


https://github.com/Opentrons/opentrons/blob/edge/CONTRIBUTING.md#opening-pull-requests

To ensure your code is reviewed quickly and thoroughly, please fill out
the sections below to the best of your ability!
-->

# Overview

Concise print statements and HS on time calculation correction.

# Test Plan

Tested on ABR robots.

# Changelog

Previously, the heater shaker temperature on time was linked to the
shaker on time. This has been corrected to reference the correct command
string. If the heatershaker is not deactivated, the on time is
calculated with the protocol end time stamp.
Changed print statements for get robot logs and abr_google_drive to make
it more obvious to the user if there is an error or not.

# Review requests

<!--
Describe any requests for your reviewers here.
-->

# Risk assessment

<!--
Carefully go over your pull request and look at the other parts of the
codebase it may affect. Look for the possibility, even if you think it's
small, that your change may affect some other part of the system - for
instance, changing return tip behavior in protocol may also change the
behavior of labware calibration.

Identify the other parts of the system your codebase may affect, so that
in addition to your own review and testing, other people who may not
have the system internalized as much as you can focus their attention
and testing there.
-->