fix(app): liquid detail fixed precision
There were a couple places remaining where we presented liquid volumes
but didn't have fixed precision, so we could have a pretty excessive
number of decimals. Now those are all toFixed(1)'d.

Also, there was an extra "total volume" label in some little
grey-background-roundrect text objs in the liquid detail map view that
would cause the background to clip behind the map. the label's not in
the designs, so I pulled it.

Closes RSQ-78