fix(api): skip command key hash generation only if command intent is SETUP (#15020)
Fixes RQA-2640

# Overview

In a previous PR we made a change from skipping hash generation if
`CommandCreate.intent == CommandIntent.SETUP` to if
`CommandCreate.intent != CommandIntent.PROTOCOL`. But a command can have
a null intent when it is a protocol command; we only expect setup
commands to have an explicit SETUP intent. So we were accidentally
skipping key hash generation for python protocols, which was resulting
in the app not being able to match run commands with their analysis
counterparts.

This PR fixes that and also adds an integration test so we don't
accidentally break this functionality again.

# Review requests

- want to make sure that there wasn't an important reason for making the
original change that I'm missing. But since all tests are still passing
I'm guessing there wasn't any functional change that required a change
in the original `if` statement.

# Risk assessment

Low. Bug fix.