docs(api): update descriptions of partial tip pickup behavior for 7.3 (#15003)
Update partial tip pickup docs to reflect the latest behavior, as added
in #15009 and to be released in robot software 7.3.

Addresses RTC-440.

[Sandbox](http://sandbox.docs.opentrons.com/docs-partial-pickup-channel-1/v2/)

- Removed "Column 12" and "Column 1" headers from Partial Tip Pickup
page. Removed tip-tracking code for Column 1 as no longer needed..
- Reworded note in `configure_nozzle_layout` API reference entry.

Already spoke with @CaseyBatten to get an idea of what needed to be
included here and what will go in separate documentation.

- <s>Should we actually put this straight into `edge` and deploy ASAP,
or can it wait a month and roll out with 2.18 docs (current plan)?</s>
PAPI 2.18 deploy will be timed with robot stack 7.3 release, which this
aligns with.

none