fix(app,api,robot-server): remove old sim flag (#15018)
Remove the disableFastProtocolUpload flag.

This flag makes the simulation stack, when simulating older protocols,
use a hardware controller built in simulation mode and a full standard
protocol context and/or engine. This is... not something we've
significantly tested and has a lot of surprising problems, and at this
point it doesn't bring much value. So remove the feature flag.

The old behavior can still be accessed via opentrons_simulate, which
will only use the simulating hardware controller when simulating
pre-engine python and json protocols.

Closes EXEC-360