fix(api): pass the calibrated jaw max offset value when we resetting instrument (#15032)
<!--
Thanks for taking the time to open a pull request! Please make sure
you've read the "Opening Pull Requests" section of our Contributing
Guide:


https://github.com/Opentrons/opentrons/blob/edge/CONTRIBUTING.md#opening-pull-requests

To ensure your code is reviewed quickly and thoroughly, please fill out
the sections below to the best of your ability!
-->

# Overview
We really only need to calibrate the gripper's max jaw offset once every
time we attach a new gripper. This means that when we call
`ot3api.reset()`, which is pretty often because it's called by
`ot3api.stop()`, we should just reuse the calibrated value so we don't
see the jaw opening and closing all the time.