fix(app-shell): monitor usb only once
We've observed a crash in the app node side that occurs inside
usb-detection likely to do with some surprisingly-shaped descriptors,
and those descriptors are polled only (a) on request or (b) when you
initialize a webusb device. So let's remove the webusb device
initialization and see if that helps.

Also, the webusb connect listeners are as far as I can tell completely
duplicative of the listeners in system-info/usb-devices, I wonder if
this was a merge conflict?

Closes RQA-2641