feat(robot-server): show module compatibility
The robot server's /modules endpoint now has a compatibleWithRobot field
that shows whether the connected module is considered compatible with
the robot it's running on.

That compatibility determination is based on the deck name for
now (module model definitions list what deck names they're compatible
with) though we might want to change it to the reverse, listing in the
deck configuration which modules are compatible with this deck.

Closes RSQ-6