feat(robot-server, app): add a new endpoint for fast-fetching all run commands (#15031)
# Overview

**Robot server changes:**

Adds a new endpoint- `GET /runs/:run_id/commandsAsPreSerializedList` to
the run commands router. This endpoint returns a list of pre-serialized
commands (that are valid json objects) of a finished run.

This endpoint is a much faster alternative to the `GET
/runs/:run_id/commands` endpoint when fetching all commands of a
completed run.

Also adds notification publishing when pre-serialized commands become
available for a run.

**App changes**

closes RQA-2645 and RQA-2647

# Risk assessment

Back-end: New endpoint so impact on existing code is close to none.
App: Medium. Fixes issues in existing behavior of handling historical runs.

---------

Co-authored-by: ncdiehl11 <ncdiehl11@gmail.com>
Co-authored-by: ncdiehl11 <ndiehl@opentrons.com>