refactor(robot-sever,api): make ChangeNotifier thread safe
We use ChangeNotifier as a wrapper around asyncio.Event, which lets the robot-server be notified of
changes within PE. This is sufficient when the notifier and the waiter operate in the same thread,
but is not inherently multi-thread safe. We can extend Event to make it thread safe by utilizing
asyncio's internal event loop call_soon_threadsafe method.