fix(app, app-shell): block HTTP refetching until successful MQTT subscription
Currently, the desktop app/ODD attempt to address the "missed update problem" in the following way:
while we subscribe to a topic, we simultaneously GET whatever equivalent HTTP resource we just
subscribed to. However, there's definitely a world (albeit a very small one) in which we receive the
HTTP response, a server update occurs, the server publishes, and then we successfully subscribe to a
topic. In this world, we've missed the update event. Solve for this by simply blocking the initial
HTTP GET until we subscribe. While the subscribe handshake could theoretically take a maximum of 2
seconds (at which point we forcefully timeout the subscribe action and fallback to polling), in
practice it's more like 250ms. We already handle failed connections and don't go through this
handshake if we can't connect to the client to begin with, so the "wait 2 second until sub failure"
scenario shouldn't realistically happen. This bug was a product of initially using retained messages
during prototyping, however we removed retained messaging by MQTT launch.