fix(app, react-api-client): handle failed analysis for RTP protocol on ODD
RTP protocols introduce a new scenario where an RTP protocol can be sent to a Flex, analyze
successfully with default values, and then have parameters input that will cause a failed analysis.
To handle this, the ProtocolSetup on the ODD should promp the user with a failed analysis modal with
call to action return to protocol setup to reselect parameters and start a new run. Keeping
consistent with our current behavior for protocols that fail analysis, the user will still have the
option to dismiss the re-parameterization CTA and proceed with the run. In addition, we need to
modify the ProtocolCard on ODD to search not only for the most recent analysis, but the most recent
successful analysis, should one exist.