chore(api): allow specifying module models to sim
The simulator setup files let you specify modules, but you can only do
it by specifying their types (i.e. magdeck, tempdeck) and not their
modules (i.e. magneticModuleV1, temperatureModuleV2). This hasn't really
mattered up til now because nothing cares, but RSQ-6 will have both
desktop and ODD looking pretty hard at which generation of module is
present on a flex, so we need a way to specify.

This adds a model key to the module section of the simulator setup (and
sets the test-flex.json and test.json up with it) so that you can
specify module models.

Closes RSQ-6