feat(app): add ODD modal for incompatible modules
This is a blocking modal takeover that pops when you have incompatible
modules (as determined by the machine) connected to your flex, and goes
away when you remove them.

This also adds a test id to the modal and app root elements (it's just
the id again) because testing-library/react doesn't offer a way to
search just by id, and it's nice to be able to test that your component
is hanging off the right root.

Closes RSQ-6